+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - #69 - added front for the Role
 - #56 - transferring the mapper from the controller to the service
 - #23 - added role caching
+- #49 - entities number check after update tests
 
 ## [1.0.0] - 30.07.2023
 - #1 - added working schedule entity